@@ -11,30 +11,30 @@
 	},
 	"private": true,
 	"dependencies": {
-		"@angular/animations": "~13.4.0",
+		"@angular/animations": "^14.3.0",
 		"@angular/cdk": "^13.3.9",
-		"@angular/common": "~13.4.0",
-		"@angular/compiler": "~13.4.0",
-		"@angular/core": "~13.4.0",
-		"@angular/forms": "~13.4.0",
+		"@angular/common": "^14.3.0",
+		"@angular/compiler": "^14.3.0",
+		"@angular/core": "^14.3.0",
+		"@angular/forms": "^14.3.0",
 		"@angular/material": "^13.3.9",
-		"@angular/platform-browser": "~13.4.0",
-		"@angular/platform-browser-dynamic": "~13.4.0",
-		"@angular/router": "~13.4.0",
+		"@angular/platform-browser": "^14.3.0",
+		"@angular/platform-browser-dynamic": "^14.3.0",
+		"@angular/router": "^14.3.0",
 		"jwt-decode": "^3.1.2",
 		"rxjs": "~6.6.0",
 		"tslib": "^2.0.0",
 		"zone.js": "~0.11.8"
 	},
 	"devDependencies": {
-		"@angular-devkit/build-angular": "~13.3.11",
+		"@angular-devkit/build-angular": "^14.2.13",
 		"@angular-eslint/builder": "1.0.0",
 		"@angular-eslint/eslint-plugin": "1.0.0",
 		"@angular-eslint/eslint-plugin-template": "1.0.0",
 		"@angular-eslint/schematics": "1.0.0",
 		"@angular-eslint/template-parser": "1.0.0",
-		"@angular/cli": "~13.3.11",
-		"@angular/compiler-cli": "~13.4.0",
+		"@angular/cli": "^14.2.13",
+		"@angular/compiler-cli": "^14.3.0",
 		"@types/jasmine": "~3.6.0",
 		"@types/node": "^12.11.1",
 		"@typescript-eslint/eslint-plugin": "4.3.0",
@@ -57,7 +57,6 @@
 		"prettier": "^2.2.1",
 		"pretty-quick": "^3.1.0",
 		"ts-node": "~8.3.0",
-		"tslint": "~6.1.0",
 		"typescript": "~4.6.4"
 	},
 	"husky": {
